requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bertopic==0.13.0
bleach==5.0.1
boto3==1.26.53
botocore==1.29.53
bpemb==0.3.4
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
conllu==4.5.2
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.32
datasets==2.8.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
flair @ git+https://github.com/flairNLP/flair@15776018b150a3b8500e64a3a474e3c52674170c
flatbuffers==23.1.4
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
fst-pso==1.8.1
ftfy==6.1.1
future==0.18.3
FuzzyTM==2.0.5
gast==0.4.0
gdown==4.4.0
gensim==4.3.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
hdbscan==0.8.29
hnswlib==0.6.2
huggingface-hub==0.10.0
humanfriendly==10.0
hyperopt==0.2.7
idna==3.4
importlib-metadata==3.10.1
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
Janome==0.4.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lda==2.0.0
libclang==15.0.6.1
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==2.0.4
more-itertools==9.0.0
mpld3==0.3
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==3.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pbr==3.1.1
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.12.0
pluggy==1.0.0
pptree==3.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyFUME==0.2.25
Pygments==2.14.0
pynndescent==0.5.8
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch_revgrad==0.2.0
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
segtok==1.5.11
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
simpful==2.9.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlitedict==2.1.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-text==2.10.0
termcolor==2.2.0
terminado==0.17.1
textprep==0.0.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
top2vec @ git+https://github.com/Lotfi-AL/Top2Vec.git@2d2647612685b3471ee53e8437d09fd78d7d82d3
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.25.1
typing_extensions==4.4.0
umap-learn==0.5.3
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
Wikipedia-API==0.5.8
wordcloud==1.8.2.2
wrapt==1.14.1
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
zipp==3.11.0
